Rename test class. Add more tests for various bufferedimage types
